refactor: renaming util.js to utils.js
